feat: add cli wrapper
Major change to allow the bot to be distributed as a pip package.

app.py has been removed and replaced with a CLI utility called
'rybo'.

Existing video generation, TTS and Reddit post processing logic
has been left as is, however now sits inside the CLI.

Removed auth.py, as credentials are now passed via environment
variables or a YAML configuration file. The default expected
location of the configuration file is `~/rybo.yaml`. A
sample configuration file has been included in the codebase.

ArgParse parameters updated to allow loading of configuration
from the command line, a configuration file, or environment
variables.

The order of precendence for loading configuration options is:

1. Configuration file
2. Environment variables.
3. CLI parameters.

Configuration options provided as CLI parameters will override
the same option provided as an environment variable, and the
environment variable will override the same option provided in
the configuration file.

Replaced all calls to "print" with module specific loggers to
ensure consistent output and facilitate configurable logging
behaviour.

Updated existing GitHub workflows to reflect the new expected
environment variable names.

Updated the readme with instructions on how to use the new CLI.

BREAKING CHANGE: app.py no longer exists, use the `rybo` command.

refs:
  - resolves #82
