+++ .dockerignore
@@ -1,5 +1,4 @@
 final/*
-config/auth.py
 .env
 client_secret.json
 cookies.json
