--- config/auth-example.py
@@ -1,18 +1,0 @@
-"""Authentication credentials used by the bot."""
-
-# FIXME: move credentials out of file.
-
-# Reddit Praw
-praw_client_id = "xxxxxx"  # noqa: S105
-praw_client_secret = "xxxxxx"  # noqa: S105
-praw_user_agent = "xxxxxx"
-praw_username = "xxxxxx"
-praw_password = "xxxxxx"  # noqa: S105
-
-# Amazon Polly
-aws_access_key_id = "xxxxxx"
-aws_secret_access_key = "xxxxxx"  # noqa: S105
-
-# Rumble
-rumble_username = "xxxxxx"
-rumble_password = "xxxxxx"  # noqa: S105
