+++ requirements-dev.txt
@@ -4,7 +4,7 @@ flake8-bandit==4.1.1
 flake8-black==0.3.6
 flake8-bugbear==23.3.23 
 flake8-docstrings==1.7.0
-flake8-isort==4.0.0
+#flake8-isort==6.0.0
 flake8-pyproject==1.2.3 
 flake8-secure-coding-standard==1.4.0
 mypy==1.2.0
