+++ src/rybo/cli/__init__.py
@@ -1,0 +1,1 @@
+"""Rybo command line utility."""
