+++ src/rybo/comments/__init__.py
@@ -1,0 +1,1 @@
+"""Reddit comment processors."""
