+++ src/rybo/publish/__init__.py
@@ -1,0 +1,1 @@
+"""Video publishers."""
