+++ src/rybo/reddit/__init__.py
@@ -1,0 +1,1 @@
+"""Reddit post processors."""
