+++ src/rybo/speech/__init__.py
@@ -1,0 +1,1 @@
+"""Text to Speech processors."""
