+++ src/rybo/video_generation/__init_-.py
@@ -1,0 +1,1 @@
+"""Video processors."""
